pyproject.toml
[project]
name = "axonius-automation-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Yaronmd",email = "yaronmord@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-playwright (>=0.7.0,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
